Made the Files in the Cloud match those from the Database

+++ src/back-end/services/CloudService.ts
@@ -5,7 +5,6 @@ import { Request } from 'express';
 import path from 'path';
 import * as fs from 'fs';
 import Formidable from 'formidable/Formidable.js';
-import File from '../models/File.js';
 import FileService from './FileService.js';
 import UserController from '../controllers/UserController.js';
 
@@ -13,7 +12,7 @@ const scriptDirpath: string = process.cwd();
 const cloudDirname: string = 'Cloud';
 const cloudDirpath : string = path.join(scriptDirpath, cloudDirname);
 
-var cloudFiles : FileData[] = []; // TODO: Remove
+var cloudFiles : FileData[] = []; // TODO: Remove. We only need the Cloud Directory content and the Metadata from the Database
 
 class CloudService {
 
@@ -33,34 +32,49 @@ class CloudService {
         this.syncCloudFiles();
     }
 
-    static syncCloudFiles() : void {
+    static async syncCloudFiles() : Promise<void> {
         cloudFiles = cloudFiles.filter(file => {
             if (file.filepath) {
               return fs.existsSync(file.filepath);
             }
             return false;
-          });
+        });
 
-        try {
-            const files : string[] = fs.readdirSync(cloudDirpath);
-            files.forEach(file => {
-                const fileId = this.getFileHash(file);
-                if(this.isFileRegistered(fileId)) { return; }
-                const filePath = path.join(cloudDirpath, file);
-                const fileData = fs.readFileSync(filePath);
-                const newFile : FileData = {
-                    id: fileId, 
-                    // TODO: Replace placeholder values with actual values
-                    owner_id: "",
-                    parent_id: null,
-                    is_folder: false,
-                    size: 1,
-                    //
-                    filename: file, 
-                    filepath: filePath, 
+        if(!UserController.currentUser) {
+            return;
+        }
+
+        try{        
+            const filenames : string[] = fs.readdirSync(cloudDirpath);
+            for(let i = 0; i < filenames.length; ++i) {
+                const filename : string = filenames[i];
+                const filepath : string = path.join(cloudDirpath, filename);
+                const isRegistered : boolean = this.isFileRegistered(filename);
+                const file = await FileService.getFileByName(filename);
+                if(!file) {
+                    console.log("File isn't in Database. Removing it");
+                    await fs.promises.rm(filepath).then(() =>{
+                        cloudFiles = cloudFiles.filter(cloudFile => {
+                            return cloudFile.filename != filename;
+                        });
+                    }, (err) => {
+                        console.error(`Couldn't remove ${filename}. `, err);
+                    })
+                    continue;
+                }
+                if(isRegistered) { return; }
+                
+                const newFileData : FileData = {
+                    id: file.id, 
+                    owner_id: file.owner_id,
+                    parent_id: file.parent_id,
+                    is_folder: file.is_folder,
+                    size: file.size,
+                    filename: filename, 
+                    filepath: path.join(cloudDirpath, filename), 
                 };
-                cloudFiles.push(newFile);
-            });   
+                cloudFiles.push(newFileData);
+            }
         } catch(err) {
             console.error(err);
         }
@@ -77,50 +91,51 @@ class CloudService {
         }
 
         if(!UserController.currentUser) {
-            console.error(`Current user is undefined`);
-            return;
+            throw new Error("Current user is undefined");
         }
-        const userId : string = UserController.currentUser.dataValues.id;
 
         if(files.multipleFiles) {
-        for(const file of files.multipleFiles) {
-            if(file.size <= 0) { continue; }
-            let newFileData : FileData;
-            if(file.originalFilename) {
-                newFileData  = {
-                    id: this.getFileHash(file.originalFilename),
-                    owner_id: userId,
-                    parent_id: null, // TODO: Get it from the Frontend (current folder)
-                    is_folder: false, // TODO: Handle Folder creation
-                    size: fs.statSync(file.filepath).size,
-                    filename: file.originalFilename, 
-                    filepath: path.join(cloudDirpath, file.originalFilename), // TODO: Maybe remove (not needed)
-                };
-                if(newFileData.filename == '') { continue; }
-                if(newFileData.id && this.isFileRegistered(newFileData.id)) { continue; }
+            for(const file of files.multipleFiles) {
+                if(file.size <= 0) { continue; }
+                let newFileData : FileData;
+                if(file.originalFilename) {
+                    newFileData  = {
+                        id: this.getFileHash(file.originalFilename),
+                        owner_id: UserController.currentUser.id,
+                        parent_id: null, // TODO: Get it from the Frontend (current folder)
+                        is_folder: false, // TODO: Handle Folder creation
+                        size: fs.statSync(file.filepath).size,
+                        filename: file.originalFilename, 
+                        filepath: path.join(cloudDirpath, file.originalFilename), // TODO: Maybe remove (not needed)
+                    };
+                    if(newFileData.filename == '' || this.isFileRegistered(newFileData.filename)) { continue; }
 
-                try{
-                    await FileService.createFile(newFileData);
-                }
-                catch(err) {
-                    console.error(err);
-                    continue;
-                }
-
-                // Write the File in the Cloud
-                if(newFileData.filepath) {
-                    try {
-                        let data : Buffer = await fs.promises.readFile(file.filepath);
-                        await fs.promises.writeFile(newFileData.filepath, data);
-                    } catch(err) {
+                    try{
+                        const newFile = await FileService.createFile(newFileData);
+                        if(!newFile) {
+                            throw new Error("Couldn't create new File in the Cloud");
+                        }
+                        newFileData.id = newFile.id;
+                    }
+                    catch(err) {
                         console.error(err);
                         continue;
                     }
+
+                    // Write the File in the Cloud
+                    if(newFileData.filepath) {
+                        try {
+                            let data : Buffer = await fs.promises.readFile(file.filepath);
+                            await fs.promises.writeFile(newFileData.filepath, data);
+                        } catch(err) {
+                            console.error(err);
+                            continue;
+                        }
+                    }
+                    cloudFiles.push(newFileData);
                 }
-                cloudFiles.push(newFileData);
+                
             }
-            
-        }
         }
     }
 
@@ -141,9 +156,13 @@ class CloudService {
         const fileIndex = Number(req.params.id);
         const file : FileData | undefined = cloudFiles.at(fileIndex);
         if(file && file.filepath && file.id != undefined) {
-
-            // Make this work:
-            //FileService.deleteFile(file.id);
+            try{
+                await FileService.deleteFile(file.id);
+            }
+            catch(err) {
+                console.error(`Couldn't delete ${file.filename} from Database. `, err);
+                return;
+            }
 
             await fs.promises.rm(file.filepath).then(() =>{
                 cloudFiles.splice(fileIndex, 1);
@@ -171,31 +190,15 @@ class CloudService {
         }
     }
 
-    static async renameFile(req: Request) : Promise<void> {
-        const fileIndex = Number(req.params.id);
-        const file : FileData | undefined = cloudFiles.at(fileIndex);
-        if(file && file.filepath && file.id != undefined) {
-
-            // Make this work:
-            //FileService.deleteFile(file.id);
-
-            await fs.promises.rm(file.filepath).then(() =>{
-                cloudFiles.splice(fileIndex, 1);
-            }, (err) => {
-                console.error(`Couldn't remove ${file.filename}. `, err);
-            })
-        }
-    }
-
     static getFileHash(filename: string) : string {
         const hash = crypto.createHash('sha256');
         hash.update(filename);
         return hash.digest('hex');
     }
 
-    static isFileRegistered(id : string) : boolean {
+    static isFileRegistered(filename : string) : boolean {
         for(const cloudFile of cloudFiles) {
-            if(cloudFile.id == id) {
+            if(cloudFile.filename == filename) {
                 return true;
             }
         }
@@ -203,8 +206,9 @@ class CloudService {
     }
 
     static getPageHtml() : string {
+        const userText : string = UserController.currentUser ? `of ${UserController.currentUser.username}` : "";
         return `
-        <h2>Personal Cloud</h2>
+        <h2>Personal Cloud ${userText}</h2>
         ${this.getCloudFilesHtml()}
         <form action="/files/upload" enctype="multipart/form-data" method="post">
           <div>Files to upload: <input id="uploadFiles" type="file" name="multipleFiles" multiple="multiple" /></div>
